+++ src/main/java/com/masiv/movies/service/ReportService.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.masiv.movies.models.Function;
 
 public interface ReportService {
-	public List<Function> getTopFunctionByTicketsSold(Long top);
+	public List<Function> getTopFunctionByTicketsSold(Integer top);
 }
